Move _package_version to travertino
