+++ changes/3183.misc.rst
@@ -1,0 +1,1 @@
+Move ``_package_version`` to travertino
